Update to 2.0.0.0 for ModularMods fork
