+++ gradle.properties
@@ -3,13 +3,13 @@ org.gradle.jvmargs=-Xmx4G
 
 # Fabric Properties
 	# check these on https://fabricmc.net/develop
-	minecraft_version=1.19
-	loader_version=0.14.8
+	minecraft_version=1.19.2
+	loader_version=0.14.10
 
 # Mod Properties
-	mod_version = 1.19-Fabric-1.0.0.0
-	maven_group = com.timlee9024.mcgltf.example
+	mod_version = 1.19.2-Fabric-2.0.0.0
+	maven_group = com.modularmods.mcgltf.example
 	archives_base_name = MCglTF-Example
 
 # Dependencies
-	fabric_version=0.58.0+1.19
+	fabric_version=0.64.0+1.19.2
